Classified inline and normal mir patterns
