+++ crates/rpl_patterns/src/cve_2018_20996_deprecated.rs
