+++ crates/rpl_patterns/src/inline/cve_2018_20992.rs
