+++ crates/rpl_patterns/src/inline/mod.rs
@@ -1,0 +1,12 @@
+pub(crate) mod cve_2018_20992;
+pub(crate) mod cve_2018_21000;
+pub(crate) mod cve_2019_15548;
+pub(crate) mod cve_2019_16138;
+pub(crate) mod cve_2020_25016;
+pub(crate) mod cve_2020_35873;
+pub(crate) mod cve_2020_35881;
+pub(crate) mod cve_2020_35888;
+pub(crate) mod cve_2020_35892_3;
+pub(crate) mod cve_2020_35898_9;
+pub(crate) mod cve_2021_25904;
+pub(crate) mod cve_2021_29941_2;
