+++ crates/rpl_patterns/src/normal/cve_2019_16138.rs
@@ -57,17 +57,6 @@ impl<'tcx> Visitor<'tcx> for CheckFnCtxt<'_, 'tcx> {
                     .dcx()
                     .emit_err(crate::errors::SetLenUninitialized { vec, set_len });
             }
-            let pattern_cast = pattern_set_len_uninitialized_inlined(self.pcx);
-            for matches in
-                CheckMirCtxt::new(self.tcx, self.pcx, body, pattern_cast.pattern, pattern_cast.fn_pat).check()
-            {
-                let vec = matches[pattern_cast.vec].span_no_inline(body);
-                let set_len = matches[pattern_cast.set_len].span_no_inline(body);
-                debug!(?vec, ?set_len);
-                self.tcx
-                    .dcx()
-                    .emit_err(crate::errors::SetLenUninitialized { vec, set_len });
-            }
         }
         intravisit::walk_fn(self, kind, decl, body_id, def_id);
     }
@@ -92,39 +81,6 @@ fn pattern_set_len_uninitialized(pcx: PatCtxt<'_>) -> Pattern<'_> {
             let vec_ref: &mut std::vec::Vec<$T> = &mut vec;
             #[export(set_len)]
             _ = std::vec::Vec::set_len(move vec_ref, _);
-        }
-    };
-    let fn_pat = pattern.fns.get_fn_pat(Symbol::intern("pattern")).unwrap();
-
-    Pattern {
-        pattern,
-        fn_pat,
-        vec,
-        set_len,
-    }
-}
-
-#[rpl_macros::pattern_def]
-fn pattern_set_len_uninitialized_inlined(pcx: PatCtxt<'_>) -> Pattern<'_> {
-    let vec;
-    let set_len;
-    let pattern = rpl! {
-        #[meta($T:ty)]
-        fn $pattern (..) -> _ = mir! {
-            // let vec: std::vec::Vec<$T> = std::vec::Vec::with_capacity(_);
-            let raw_vec_inner: alloc::raw_vec::RawVecInner = alloc::raw_vec::RawVecInner::with_capacity_in(_, _, _);
-            let raw_vec: alloc::raw_vec::RawVec<$T> = alloc::raw_vec::RawVec::<$T> {
-                inner: move raw_vec_inner,
-                _marker: const std::marker::PhantomData::<$T>
-            };
-            #[export(vec)]
-            let vec: std::vec::Vec<$T> = std::vec::Vec::<$T> { buf: move raw_vec, len: const 0_usize };
-
-            let vec_ref: &mut std::vec::Vec<$T> = &mut vec;
-
-            #[export(set_len)]
-            // _ = std::vec::Vec::set_len(move vec_ref, _);
-            ((*vec_ref).len) = _;
         }
     };
     let fn_pat = pattern.fns.get_fn_pat(Symbol::intern("pattern")).unwrap();
