+++ crates/rpl_patterns/src/normal/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod cve_2019_15548;
+pub(crate) mod cve_2019_16138;
+pub(crate) mod cve_2021_25904;
+pub(crate) mod cve_2021_27376;
+pub(crate) mod cve_2021_29941_2;
+pub(crate) mod cve_2022_23639;
